Use feature flag. Add additional scenario routes
