+++ src/platform/site-wide/mega-menu/mega-menu-link-data-for-authenticated-users.json
@@ -5,6 +5,7 @@
     },
     {
         "title": "My Health",
-        "href": "/my-health-account-validation/"
+        "href": "https://www.myhealth.va.gov/mhv-portal-web/home",
+        "target": "_blank"
     }
 ]
